pyproject.toml
[project]
name = "wishdex-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.0.0",
    "scrapy>=2.8.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "prettytable>=3.0.0",
    "beautifulsoup4>=4.9.0",
    "requests>=2.25.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.2.1",
    "anthropic>=0.73.0",
    "google-generativeai>=0.8.5",
    "google-genai>=1.51.0",
    "jsonref>=1.1.0",
    "instructor>=1.13.0",
    "playwright>=1.57.0",
    "fastapi[standard]>=0.128.0",
    "selenium>=4.39.0",
    "webdriver-manager>=4.0.2",
    "chromedriver-autoinstaller>=0.6.4",
    "undetected-chromedriver>=3.5.5",
    "fake-useragent>=2.2.0",
    "curl-cffi>=0.14.0",
    "asyncpg>=0.31.0",
    "httpx>=0.28.1",
    "pylint>=4.0.4",
    "flake8>=7.3.0",
    "black>=26.1.0",
    "isort>=7.0.0",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "ruff>=0.14.1",
    "pytest-snapshot",
]

[tool.pytest.ini_options]
markers = [
    "e2e: end-to-end tests",
    "unit: unit tests"
]
